feat: refactor styles for authentication pages; consolidate CSS into auth.module.css
